+++ mycodo/inputs/ttn_data_storage.py
@@ -68,7 +68,7 @@ INPUT_INFORMATION = {
     ],
 
     'dependencies_module': [
-        ('pip-pypi', 'requests', 'requests==2.25.1'),
+        ('pip-pypi', 'requests', 'requests==2.31.0'),
     ],
 
     'custom_options': [